Allow long keyids and key fingerprints (prefered).
The key owner is now able to pass his/her keys as short or long keys,
with or without trailing 0x, as well as full fingerprints (formatted or
not).

These formats can also be use freely in the input file; if fingerprints
are present (e.g. if the file was generated with gpg --fingerprint) they
take precedence.

Closes: #736963