Add an option to add non-exportable signatures in ~/.gnupg.
In addition to caff's signatures in its own GnuPGHOME.  This can be
useful when the recipient forgets to upload the signatures caff sent (or
if they non-exportable as well), as it gives a way to keep track of
which UIDs were verified.  However, note that local signatures will not
be deleted once the recipient does the upload and the signer refreshes
her keyring.

The new configuration option $CONFIG{'also-lsign-in-gnupghome'} can take
the values 'yes', 'ask' or 'no'.  (Default: 'no').
  * If not 'no' and if $CONFIG{'gpg-sign-type'} contains "l" (i.e., caff
    sends local signatures), each (local) signature is merely exported
    from caff's own GnuPGHOME to the user's.
  * If 'yes' each UID signed in caff's own GnuPGHOME gets automatically
    locally signed in the user's; this requires a working gpg-agent(1).
  * If 'ask', the user is prompted for which UIDs to locally sign.

Closes: #680136.