More robust way to prune keys (using UID hashes).
Since we're now doing the listing before-hand we don't have to parse the
output here.  I believe the refactored pruning algorithm is simpler, and
also more robust as we can call delete_signatures with an UID hash (as
opposed to an index).  It's also shorter and requires less interaction
with the GnuPG shell as UIDs don't have to be unselected (and when we
need to, 'uid 0' does unselect them all).