Bug #735363 was fixed upstream; removing customing fix for #735536.
The situation reported in #735536 (and #735363) no longer fails as of
GnuPG 1.4.16-1.  However, while the ad-hoc fix introduced 1.4.15 did
solve the issue by forcing gpg to create a trustdb, it also made
GnuPG::Interface croak:

    Use of uninitialized value $command[8] in exec at /usr/share/perl5/GnuPG/Interface.pm line 301.

(This is only visible when checking the $stderr in readwrite_gpg, for
instance by making trace2 visible.)

This is because '--import-ownertrust' shouldn't be passed to hash_init's
'extra_args', as an *option*, but to wrap_call's 'commands' instead, as
a *command*:

    my $pid = $gpg->wrap_call(commands => [ '--import-ownertrust' ], handles => $handles);

However, since the bug was fixed upstream, and as calls to gpg are
expensive, (create_trustdb() had to be called for each single UID of
each key in order to fix #735536), I removed the had-hoc solution.