gpgsigs: Display the correct size of UATs.
This requires to use gpg's --attribute-file and --status-fd.  The
correct size of the attribute (and not the size of the whole subpacket)
plus a 16-byte header is displayed on the status FD.

This allows to manually split the attribute file, therefore removes the
need for a helper photo-viewer script.

Finally, I added a dependency to GnuPG::Interface which allows simpler
handling of multiple output FDs.