Make pgpring display the correct key length for DSA and Elgamal keys.
The Public-Key Packet Formats is described in RFC 4880, section 5.5.2:

    https://tools.ietf.org/html/rfc4880#section-5.5.2

The meaning of the first 6 octets is identical for all algorithms.  The
public-key algorithm itself is given by the 6th byte as per RFC 4880
section 9.1.  Then follows a number of multiprecision integers (MPI)
depending on the algorithm: RSA (algorithm 1 to 3) public keys have 2
MPIs, DSA (algorithm 17) public keys have 4, and Elgamal (algorithm 16
and 20) public keys have 3.

What is commonly called "length" of the key is always the length of the
first of said MPIs (public modulus in case of RSA, prime p in case of
DSA and Elgamal).

Patch from Fabrizio Tarizzo.  Closes #602284.