Rely on the field no. only when parsing the output of --list-colons.
/usr/share/doc/gnupg/DETAILS.gz clearly specifies each field, but not
the total number of fields per selector.  (This is on purpose since it
leaves room for future extensions, such as EC in GnuPG 2.1.x.) Therefore
a robust regex may only rely on the field index, i.e., on the number of
colons ‘:’ met.

As for the last field, from the specifications it's unclear whether
there is a trailing colon or not.  Hence our regexes should deal with
both cases when there is ambiguity.