Color update and changed functionality
keyart now accepts any valid input that will uniquely identify a key in your
keyring, rather than relying on an exported key file. This identification can
be a name, email address, short keyid, long keyid, fingerprint, etc.

Updated the printed ANSI colors to give less emphasis on the blues.

Finally, changed one of the coins from '{' to '(', as printing a literal curly
brace in Python with .format() is tricky, and not worth the effort.