Much code update and many manpage update
The '-i|--id' and '-k|--keyring' options are now mutually exclusive, require at
least one argument, and one of them is required to be passed to the command.
Currently, '-k|--keyring' code is not implemented- coming soon.

Did some code changing to introduce a 'test_env()' function and moved logic
around a bit making it easier to work with. Also micro-optimized creating the
little endian array.

Updated the keyart.1 manpage to reflect the new options.