Full public key and keyring support. Change option
Rather than '-k|--keyring' for the option to read a file, '-f|--file' was used.
This allows the exported file to be either an individual exported GPG key or a
public keyring. Similarly with '-i|--id', multiple files or keyrings may be
passed to '-f|--file'. The keyart.1 manpage is updated to reflect this change.