caff: Prune keys with import-{clean,minimal} not export-{clean,minimal}.
We want gpg(1) to remove all sigs that are neither ours or self-sigs,
because it's faster than letting caff do it using delete_signatures(),
so import-clean is really what we need.

  gpg --homedir ~/.caff/gnupghome --export $CONFIG{keyid} | gpg --homedir $keydir --import-options import-minimal --import
  gpg --homedir ~/.caff/gnupghome --export $keyid         | gpg --homedir $keydir --import-options import-clean   --import

It doesn't make a difference to use import-minimal vs export-minimal,
but athough it clutters the pipe the former is faster, presumably
because gpg(1) walks through the entire keyring in the latter case.